backend/requirements.txt
fastapi
uvicorn
stockfish
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]